requirements.txt
Django==1.11.6
django-cors-headers==2.1.0
django-filter==1.1.0
django-resized==0.3.8
djangorestframework==3.7.1
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
PyMySQL==0.7.11
pytz==2017.2
